requirements.txt
asgiref==3.11.1
Django==5.2.12
djangorestframework==3.17.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.9.0
drf-yasg==1.21.7
Pillow==12.2.0
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
